\name{BKPC-package}
\alias{BKPC-package}
\alias{BKPC}
\docType{package}
\title{
Bayesian Kernel Projection Classifier
}
\description{
Bayesian kernel projection classifier is a nonlinear multicategory classifier which performs the classification of the projections of the data to the principal axes of the feature space. A Gibbs sampler is implemented to find the posterior distributions of the parameters. The main function for end users is   \code{\link{bkpc}}.
}


\details{
  \tabular{ll}{
    Package: \tab BKPC\cr
    Type: \tab Package\cr
    Version: \tab 1.0 \cr
    Date: \tab 2016-02-27\cr
    License: \tab GPL-2\cr
  }
}
\author{
K. Domijan <domijank@tcd.ie>
}

\references{
Domijan K. and Wilson S. P.:
Bayesian kernel projections for classification of high dimensional data.
Statistics and Computing, 2011, Volume 21, Issue 2, pp 203-216
}

  
\keyword{ package }
\seealso{
\pkg{kernlab}
}

